requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
pillow>=10.4.0
numpy>=1.26.4
tensorflow>=2.17.0
python-dotenv>=1.0.1
jinja2>=3.1.4